streamlit_app/requirements.txt
streamlit
pandas
numpy
requests
libpq-dev #requirement for psycopg2
psycopg2-binary
matplotlib
